// front-end/src/app/interfaces/profiledata.ts
export interface Profiledata {
    acctno: string;
    date: string;
    creationDate: string;
    amount: number;
    paymentmethod: string;
    product: string;
    deviceid: string;
    imei: string;
    finalstatus: string;
    channelid: string;
    transactionid: string;
    blacklisted: boolean,
    rulesflag: string,
    listedunder: string,
    fraudflag: string,
    fraudmessage: string,
    rulesengineaction: string

}